+++ frontend/index.html
@@ -20,6 +20,7 @@
 			crossorigin="anonymous"
 		></script>
 
+
 		<!-- SCRIPTS -->
 		<script type="module" src="./js/router.js" defer></script>
 	</head>
