+++ frontend/js/router.js
@@ -1,11 +1,11 @@
 const routes = {
 	"/": {
-	  component: "home-page",  // Home component will load HomeContent
-	  path: "../pages/components/home/Home.js",  // Dynamically load the Home component
+		component: "home-page", // Home component will load HomeContent
+		path: "../pages/HomePage.js", // Dynamically load the Home component
 	},
 	"/home": {
-	  component: "home-page",  // Home component will load HomeContent
-	  path: "../pages/components/home/Home.js",  // Dynamically load the Home component
+		component: "home-page", // Home component will load HomeContent
+		path: "../pages/components/home/Home.js", // Dynamically load the Home component
 	},
 	"/login": {
 		component: "login-page",
@@ -16,9 +16,9 @@ const routes = {
 		path: "../pages/RegisterPage.js",
 	},
 	404: {
-		component: "not-found-page",  // Use a 404 fallback component
-		path: "../pages/NotFound.js",  // 404 not found component path
-	  }
+		component: "not-found-page", // Use a 404 fallback component
+		path: "../pages/NotFound.js", // 404 not found component path
+	},
 };
 
 const redirect = (path) => {
